[BP] MB-52842: Fix intermittent failure in 'disk>RAM delete paged-out'
The test has long been racy / broken - it (incorrectly) assumed that
attempting to read a non-resident, deleted item would avoid a bg_fetch
and directly return ENOENT. This is flawed as deleted items _are_
added to the Bloom filter in full-eviction mode.

However, this test has been incorrectly passing as it did not wait for
the previous Set to be persisted to disk - which is when the key is
added to the Bloom filter. As such, if the verify_key() call executed
before flusher had completed then the first would report as passing,
as there was no key in the Bloom filter.

Fix the test to check the behaviour I assume it was attempting to
check - wait for persistance of the delete; then run compaction and
purge the tombstone so the newly-built Bloom filter doesn't include
the requested key.

(cherry picked from commit cdc3c2f29ea6bd858faea69190031d17c2879fb8)

Change-Id: I9bbe7dd32f7e23752bccf46327f4275aeab1427d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182625
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Well-Formed: Restriction Checker
Tested-by: Jim Walker <jim@couchbase.com>