MB-52842: Fix intermittent failure in 'disk>RAM delete paged-out'
The test has long been racy / broken - it (incorrectly) assumed that
attempting to read a non-resident, deleted item would avoid a bg_fetch
and directly return ENOENT. This is flawed as deleted items _are_
added to the Bloom filter in full-eviction mode.

However, this test has been incorrectly passing as it did not wait for
the previous Set to be persisted to disk - which is when the key is
added to the Bloom filter. As such, if the verify_key() call executed
before flusher had completed then the first would report as passing,
as there was no key in the Bloom filter.

Fix the test to check the behaviour I assume it was attempting to
check - wait for persistance of the delete; then run compaction and
purge the tombstone so the newly-built Bloom filter doesn't include
the requested key.

Change-Id: I5ef2479ac7e6a4670e50d0036d585854f4490ec5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/178478
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>