MB-53294: Rename Cardinality -> MetricGroup
The name "Cardinality" was used for consistency with the naming
convention established by ns_server.

However, with the introduction of Metering metrics Cardinality is now
too specific; Metering is not a cardinality, but is a new metrics
endpoint alongside the existing "high" and "low".

To better reflect this, rename Cardinality to MetricGroup so Metering
can be added. This will be done in a following patch, along with adding
the first engine-specific Metering metrics.

This may also pave the way for a future rename from "high" and "low" to
something more internally meaningful (e.g., something conveying
"important" or "frequent").

Change-Id: I8ffaebdf4306e509255007ff44c116e52bf344be
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/178602
Tested-by: James H <james.harrison@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>