MB-53359: Add uniqe error code for config-bucket
Return EConfigOnly (0x0d) for clientw which enabled xerror
to indicate that the command can't be executed in a config-only
bucket.

For clients without xerror enabled we need to use one of the
initial error codes, and given that some commands may return
tmpfail (for instance when the server is low on memory) use
"Not Supported" instead.

Change-Id: I1e6c4f6fcb7a01a9958d5e9d63f0dd393fdcd886
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/178870
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>