MB-53415: Reset scan status for each item during warmup scan
During warmup we yield scans after they have exceeded a certain
duration to allow other tasks to run in a timely manner. The scan
callback function using during warmup may also return early for
uninteresting items which includes prepares of durable writes and
collections system events. When these items are scanned the scan
status is not currently reset. This means that the following
sequence of events is possible (for a value-eviction bucket):

1) Warmup reaches the KeyDump phase of warmup
2) Warmup scans an item at seqno 1 which is loaded before yielding
   the scan due to the time taken which returns scan_again to the
   VBucket visitor which interprets this as a yield (scan must run
   again)
3) Warmup continues the scan and sees a prepare at seqno 2, the early
   return is hit but the scan status is not reset returning another
   scan_again to the VBucket visitor which interprets this as a stop
   for the vBucket (enough data has been loaded)
4) Get performed for any keys after seqno 2 fails as no metadata is
   present in memory for the key

Fix the issue by always setting the scan status to success at the top
of the function.

Change-Id: Icd6d278afa40316c5e7bc45d629ef76eb7934120
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179067
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>
Well-Formed: Restriction Checker