MB-53304: Enforce holding of stateLock in VBucket::queueItem [1/3]
The approach suggested in MB-53304 of ensuring that the stateLock is
held by passing a const& to it in the call hierarchy of queueItem
revealed several code paths that needed to hold that lock but did not.

This patch contains the findings of that work. Later patches will
implement the passing of the lock as a parameter.

Change-Id: Icca8771aeb839342e0884da960c62185249dbab6
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/178930
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>