MB-53304: Enforce holding of stateLock in VBucket::queueItem [2/3]
The MB-53304 suggests we ensure that the stateLock is held by passing a
const& to it in queueItem (and the call hierarchy of queueItem).

The stateLock is a `folly::SharedLock` and has two 'holder' variants.
We don't really care whether a `SharedLock::WriteHolder` or
`SharedLock::ReadHolder` is being held, as long as either one is.
Unfortunately, these two types do not implement the same base class and
so we cannot pass a `WriteHolder` to a `const ReadHolder&`.

This patch introduces a logical type-erased reference type
`SharedLockRef` which can be assigned from either of the two types of
locks and does nothing otherwise. `SharedLockRef` is a template which
can be instantiated with different 'tag types' to create a unique
type-per-locked-state. See the `VBucketStateLockRef` typedef.

A follow-up patch will add a `VBucketStateLockRef` to `queueItem` and
all of its callers (up to the point where we lock the vb state).

Change-Id: Icb8f1a174c0b97278983e3cf8d401f993f0138ad
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/178931
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>