MB-52676: Split eviction eligibility check into two stages
A later patch will track the MFU of all evictable items.

Items cannot be evicted from replica vbuckets under ephemeral, as this
would make them inconsistent with the active; as such for ephemeral
items are considered "not eligible for eviction" if the vbucket is
currently a replica.

However, the MFUs of items which _could_ be evicted still need to be
tracked, as the vbucket may be promoted to active in the future.

This requires the ability to correctly class a value as "evictable",
regardless of vbucket state.

This patch makes splits vbucket state dependent eligibility checks out
of eligibleToPageOut, so a later patch can make use of the
non-state-dependent checks.

Change-Id: Iea84e3dea41ebb260ac29a9c2dc14d50e5598435
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179219
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>