MB-53448: DCP_ADD_STREAM_FLAG_TO_LATEST should use highSeqno of collections(s) in filter
If a DCP stream is requested for a legacy (non-collections) client,
then we correctly send only the default collection (as the client
knows nothing of collections).

However, if the client specifies the DCP_ADD_STREAM_FLAG_TO_LATEST
flag to stream up to the "latest" seqno of the default collection, and
the high_seqno of the vBucket is not from the default collection, then
while the stream sends the correct data is does not send a StreamEnd
message after the last "current" default collection mutation - instead
the stream just hangs which is not expected behaviour for TO_LATEST.

The issue is in the calculation of the end seqno for the stream - it
always uses the VBucket high_seqno, and not the high seqno of the
default collection (or any other filtered DCP collection).

Fix by using the same calculation as we use for rollback - the maximum
of the collection_high_seqno for the selected collections - _default
in the case of the original bug.

Change-Id: I7b966b6b4c73e0e82ef548c913132d0b3bbfd913
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179122
Reviewed-by: James H <james.harrison@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>