Cleanup: remove 'polling' durability timeout mode
Timeout handling of SyncWrites was initially done (in mad-hatter) via
polling for any 'overdue' SyncWrites periodically. This was
inefficient if there was no SyncWrites which were actually timed out,
particulary with high bucket counts (see MB-39815). As such, an
event-driven implementation was added (in neo). but we kept the
previous polling mode in case we had to revert to the old method.

The event-driven method has been present (and the default) since 7.1.0
and it working correctly, so remove the old, unused polling mode to
simplify code maitenance.

Change-Id: Ic0039f2582d2e621981b6e382ed217651f93cd40
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179209
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>