MB-52676: Track the MFU value of all evictable items
Eviction seeks to identify "cold" items to evict - items which have
been used least/least recently, as determined by its Most Frequently
Used counter.

Previously, the distribution of MFU values was learned progressively as
eviction ran.

However, future work may change from evicting some percentage of the
items in _each_ vbucket, to evicting some percentage of all items -
even if that leads to uneven vbuckets.

This would better keep the hottest items in the whole bucket resident
e.g., a very cold vbucket may have everything evicted, rather than
having the same percentage resident as a much hotter vbucket.

To support this, it is required that a distribution of MFUs across the
entire bucket could be approximated. Learning this incrementally would
be heavily influenced by the order vbuckets are visited in, and may be
quite inaccurate until paging has visited every vb (at which point it
is too late).

As as step towards this goal, this patch tracks the MFU of all
evictable items in a vbucket "up front"; any action which may change
the evictability of an item, or its MFU, now updates a histogram.

Change-Id: I56c1a86b9ca9cfd927692f90b4b8f4acc56586b9
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179220
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>