Fix STItemPagerTest CheckedExecutor failures
With MFUs set probabilistically the replica items may not be
evicted in one pass. This would cause runNextTask to throw a
runtime_error when we attempt to run "Item pager no vbucket
assigned" when instead the item pager needs to run again.

By setting all items' MFU value to zero, we force 0% or 100%
eviction of a vBucket if at least one item needs to be evicted,
avoiding this failure point. The test must be tweaked slightly
to account for this.

ActiveEvictedIfReplicaEvictionInsufficient needs to be tweaked
slightly to support this change. No other callers need changes.

Change-Id: I1b6be69c3075c5afffc11ea8a6d13d9a1d03d353
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179229
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>