MB-53284: Use magma memory optimized writes in BucketQuotaChangeTest
The test BucketQuotaChangeTest.QuotaChangeDownMemoryUsageHigh fails
sporadically if magma does not free memory quickly enough on it's
background threads. Magma has a memory optimized writes mode that
flushes to disk on the calling thread rather than a background thread to
avoid having extra copies of the items. Use it here to stop the test
from failing sporadically.

Change-Id: I19c69b2957a6877ebacce6aad2a998f1c6fcedcf
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179269
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>