MB-53055: Fix Checkpoint::isEmptyByExpel() semantic
As part of MB-39344 we have fixed the semantic of
Checkpoint::getHighSeqno() from (erroneously) accounting for seqnos of
also meta-items to accounting only seqnos of non-meta items.

But we missed to fix Checkpoint::isEmptyByExpel() to. In the old
definition of Checkpoint::highSeqno a checkpoint is empty only if it
shapes like [e cs). In the new definition of C::highSeqno also a
checkpoint like [e cs vbs) is empty.

That inconsistency is failing QE tests at CM::registerCursor() as we
essentially end up executing non-empty checkpoint seqno-checks on empty
checkpoints.

Change-Id: If8f7c5d28dd8433866090e97ef7b4f4012f8a274
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179284
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>