MB-50984: Default checkpoint_destruction_tasks=2
Increased from 1.
This is in preparation for removing the hard limit on the max number of
checkpoints per vbucket. Preliminary perf tests show that our
CheckpointDestroyerTask doesn't keep up with a higher number of
checkpoints. The same preliminary tests suggest that increasing to 2 is
enough for addresing the new checkpoint mem-releasing requirements.

Change-Id: I823fa9fc11e6a9068510d145e9520a23cf2bb190
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179353
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>