MB-53523: Only check snap start vs last snap end if active VB
We currently check the snap start that we send against the last snap end
to ensure that we do not have any snapshot violations. This check is not
valid for view streams which stream from replica vBuckets as we merge
disk and memory snapshots together and the first marker includes the
extent of the checkpoint items at the point of backfill and another will
be sent if more checkpoint items have been added. As such, we need to
skip this check.

Change-Id: I940ae741fcaeaf782ad8808a2d911b9c94c2e85d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179354
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>