MB-35297: Move BackfillTrackingIface out of DcpConnMap
This change is a refactoring change that forms part of the overall
change which will ensure RangeScans are capped in terms of how many
file-descriptors are in-use.

The current BackfillTrackingIface which is implemented in DcpConnMap
is used to count and limit how many DCP backfills are running, ensuring
DCP doesn't just keep opening files and consuming all file-descriptors.

RangeScans will also participate in this calculation and will need to
call into this object.

The refactoring:

1) moves the tracking out of DcpConnMap and into KVBucket

2) Rename BackfillTrackingIface as KVStoreScanTracker, fundamentally
we care about how many ScanContexts have been created by KVStore::scan
be it RangeScan or DCPBackfill.

3) Modernise the tracking code, replacing a std::mutex with
folly::Synchronised

Change-Id: I9e506fecd7ce576aa38e2c59de3c44a08e5e67ad
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179408
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>