MB-53370: Do not leak items if conn drops during vkey stats call
Validating key stats calls are split into two phases. First the item is
scheduled to be fetched from disk and ewouldblock is returned. Once the
call is made again, the loaded item is used to compute key stats.

Before this change, we used to load items into a lookups map
{cookie->item} and would only remove items from it during the second
call. If the connection is dropped, the second call is never made and
the item is never freed.

Items are now loaded into the engine-specific storage and released
automatically by the engine.

Change-Id: I2a2df40b3223e2f3e50e87c0301656db41cb5bf5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179718
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>