MB-53304: Enforce holding of stateLock in VBucket::queueItem [3/n]
MB-53304 suggests we ensure that the stateLock is held by passing a ref
to it in queueItem (and the call hierarchy of queueItem).

This change will allow the ref passed down to be either a lock on the
individual vbucket state or on all states (`vbSetMutex`).

In a future change I will change function signatures to receive this
ref and eventually pass it down to queueItem.

Change-Id: I953f9b8c00256f39d2845b330a18122b053a5b96
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179511
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>