MB-53304: Enforce holding of stateLock in VBucket::queueItem [6/n]
MB-53304 suggests we ensure that the stateLock is held by passing a ref
to it in queueItem (and the call hierarchy of queueItem).

This change ensures that we hold a lock on the VBucket state during
SyncWrite commit/abort.

In a future change I will pass this ref down to queueItem.

Change-Id: I814d6b40f75e0a0db60ff61955b9aa540469e23f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179518
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>