[BP] MB-52276: Make ActiveStream::getItemsRemaining consistent
Quantity computed as the sum of (1) the outstanding items in
checkpoints and (2) the outstanding items in the stream readyQ.

Currently (1) accounts for all items (meta + non-meta) while (2) only
for non-meta items.
By this patch we align everything to (1).

That allows also to remove the Stream::readyQ_non_meta_items counter,
now unused.

Change-Id: Id1c4e0e898472a2f068dce30272c88be0ab7459a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/194214
Well-Formed: Restriction Checker
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>