MB-52958 Add function to MagmaKVStore to return if value is compressed
Knowing whether a value is already compressed or not by kv-engine can
help Magma decide whether to do its block level compression or not in
certain scenarios.

Magma's SSTable write path allocates 5 buffers of size as large as the
value and copies the value in them. When the values are large (eg 20MB)
these copies bloat the memory consumption. One of those copies is due to
the output compression buffer. Since with large values we expect one
data block to only contain one value, Magma can skip its block level
compression if the value is already compressed. This is expected to be a
common case and helps Magma eliminate one copy.

To enable this we introduce a IsValueCompressed function that can help
in making this decision.

Linked magma change:
https://review.couchbase.org/c/magma/+/179587

Change-Id: Ic4678a764790e63a737962eada3a3278cf451a7a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179589
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>