MB-53584: Reset throttled cookies during Bucket delete
Currently we hang if we delete a Bucket when we have throttled cookies
in the system. This is because we set cookies ewouldblock state to true
when we throttle them so we must wait for them when disconnecting
connections and we stop ticking the throttle gauge then a Bucket is not
in the ready state so no notification is made to progress the
cookie/connection.

Reset the throttled cookies as we wait for clients to disconnect to
avoid processing them and avoid the Bucket deletion getting stuck.

Change-Id: I236bb725fb5329b87ca4b8fd7c3f8883e241359f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179608
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>