MB-49071: Make engine-specific storage type-safe [1/2]
Using an interface type for the engine-specific storage allows us to
avoid reinterpret_casts from void* and to use smart pointers for
automatic memory management of objects with lifetime bound by the
command context.

This will be a two-part patch. This first part swaps out the underlying
void* storage with EngineStorageIface and ensures that memory
allocations are correctly accounted for, but it does not change the
interface of the (get|set)EngineStorage methods we use in
EventuallyPersistentEngine to acccess the stored data.

I will refactor those methods and their callers to EngineStorageIface
in a follow-up patch.

Change-Id: Idadd77ded7a7be14bf7ca290f7aa29861f294442
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/163999
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>