MB-53570: Make CM::registerCursor resilient to first empty checkpoint
As part of MB-39344 we have added some sanity checks at
register-cursor. As seen in MB-53570 it turns out that some of those
checks are too strict and miss to consider that a CheckpointList like

  [disk | e:1 cs:1 x vbs:3) [memory | e:3 cs:3)

is legal on a vbucket after some state transition. See new unit test
for details.

The failure here is failing many QE tests, so it's a blocker.
This is a quick fix where we just relax the sanity checks. The existing
logic in CM::registerCursor handles the case with no further change,
proved by a specific unit test.

Some suboptimal behaviour has been spotted with regard to potential
unnesessary backfills triggered though. That will be addressed in
MB-53616.

Change-Id: I6d8152afaf30b49c4ac2fb9665c9a76acd4724f7
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179618
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>