[testapp] Fix raciness in MaxConnectionTest
Closing connections on the server happens in multiple steps,
and depending on the scheduling the client may be able to
connect to the server _before_ the actual disconnect happened
on the server. To work around the problem record the number
of idle user and system connections when starting the unit
test, and wait for the server to return to that state
before starting the next unit test.

Change-Id: I85fde9567cd3dee08ad1a0ae6c932debdbe623cd
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179707
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>