MB-54605: [testapp] Fix raciness in MaxConnectionTest
Closing connections on the server happens in multiple steps,
and depending on the scheduling the client may be able to
connect to the server _before_ the actual disconnect happened
on the server. To work around the problem record the number
of idle user and system connections when starting the unit
test, and wait for the server to return to that state
before starting the next unit test.

Change-Id: Ia8e469e7dc19216e8285735f2d3b0f9e27fab40e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182978
Well-Formed: Restriction Checker
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>