MB-53304: Enforce holding of stateLock in VBucket::queueItem [4/n]
MB-53304 suggests we ensure that the stateLock is held by passing a ref
to it in queueItem (and the call hierarchy of queueItem).

This change ensures we hold a lock on the VBucket state while updating
scopes and collections.

In a future change I will pass this ref down to queueItem.

Change-Id: Ifb0487a72fe17e3c488a63e6c9c348e91ebb87fa
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179512
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>