MB-35297: Evaluate all RangeScan continue limits
The current if/else if block is incorrect and for example means that
if both an item + byte limit are enabled, the item limit is the only
one checked, and the scan may not end at the expected point.

The updated if/else if now checks all limits for each call.

Change-Id: Ib606c164aa9060f459ffba75ca4741d94fb3f71c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179901
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>