MB-35297: Implement RangeScan throttling
The CookieIFace is updated to provide

* sendResponse (moved from server_api.cc)
* checkThrottle

checkThrottle is a new function that takes as input the current
accumulated read/write cost of a partially complete command and
checks the bucket throttle using this "pending" cost. Note that
the pending read cost of the continue is updated irrespective of
the handler's buffering.

The RangeScan read callbacks now evaluate the throttle for each
read key/item and will stop the current continuation if the
checkThrottle returns true. The pending cost of the range-scan-continue
is then fed back into the metering stats when the closing response
is processed.

Change-Id: If13156ade7c6212588e1498f3dd2f96a376d7413
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179939
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>