[serverless] serverless tests should only run on Magma
The serverless test suite tries to test features related to serverless
deployment, and serverless deployments must use the magma backend.
There is no point of trying to run all of the tests twice (on each
backend) when only one is being used.

To allow some runs on couchstore, run the tests with couchstore on
CE configurations.

Change-Id: Id71dc4515210e92ee0301914ef74feebf9556a5c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/180054
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>