MB-53597: Add task group for all tasks of all taskables
We currently do not gather stats for the NoBucket taskable but we are
adding more tasks to it for the cross-Bucket memory sharing work. These
stats may be useful in the future, so add an additional stat group that
gathers stats from all Ready Buckets (we cannot touch non-Ready Buckets
easily) and the NoBucket taskable for later use in cbcollect_info.

Change-Id: I56a00a1e92c32bc9aa39c44e1e090d98b8f8aeb9
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/180238
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>