Add an indication for the next throttle tick
Insert '{ "next_tick_us" : <number> }' in the response message
for EWouldThrottle responses. That would give the client a
hint for when we think the next tick will run on the server.

The client may use the value as a hint for how long it should
back off before retrying the operation. Note that the value
is not accurate as it is another thread which perform the
tick on the buckets and it performs the tick in sequence
bucket by bucket.

Change-Id: I0dbbcc09d2ac3ae38110ead66b2ab6e299edc186
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/180305
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: James H <james.harrison@couchbase.com>