MB-52684: Set Zstd as default for data block compression
Change default data block compression algorithm from LZ4
to Zstd set to level-1 compression. Zstd provides ~30% better
compression ratio in perf and YCSB tests. It consumes around
5% extra CPU on the write side in Magma's background flusher
threads. On the read side Zstd is slower than LZ4 but since
this is only for the data blocks in the sequence index which are
not cached and the get has to go to disk, we did not see any
noticable change in read latency.

Zstd is set to compression level-1. We ran YCSB and magma 1%
dgm perf tests with level 1, 3 and 5. The compression ratio did
not improve substantially as the level was increased (level-5 had
~5% additional compression ratio) but the CPU usage was higher
by 20-25% over level-1. Level-1 does not use much more CPU than
LZ4 during compression and gives a ~30% boost in compression ratio.

Change-Id: I5670b8b6c05fdd61c037063d6109c4aecdce80a9
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/180341
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>