Don't log missing privs for connection self stat
checkPrivilege will log if the privilege is missing and inject
an UUID in the response with the matching log entry.

All connections should be allowed to look at itself, so we
don't error out, but limits the connection to itself
unless it requested a given connection. It feels a
bit weird to generate log messages in this case.

Change-Id: I13919a3f1d8e5ca5c167180933df606759bcd631
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/180352
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>