MB-53898: Ensure we finish timing getMulti at the same time
Currently in getMultiCallback() we set the end time of when
CouchKVStore::fetchDoc() completed, at different times. Due to us
calling steady_clock::now() every time we added a time delta to
readTimeHisto. This means we will have a very slightly skewed values in
readTimeHisto when we have a large number of documents that are being
read. Instead we should just call steady_clock::now() once after we've
called CouchKVStore::fetchDoc().

Change-Id: Ia07c8f83fae1e2480737999895132e5c9525d2bf
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/180652
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>