MB-46740: CM::getItemsForCursor() allows to specify a byte-limit
The function already accepts 'approxNumItemsLimit' in input.
The new 'approxBytesLimit' arg allows the user to specify a further
bytes-limit.
The two limits are orthogonal. The function triggers whatever we hit
first.

The semantic of of the new arg is the same of the pre-existing limit.
Apart from some exception, in general the function always moves the
cursor at checkpoint boundaries for generating consistent snapshots.
So, the function moves the cursor to a checkpoint boundary that is
greater/equal to the lower of any limit specified by the caller.

Feature added with full test coverage but not used yet in he prod code.
We'll start using the new API in DCP in follow-up patches.

Change-Id: Idb88ab6d4682027311418a9fcde691ab949a5ed1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/180584
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>