MB-53922, MB-54295: Remove ~StoredValueProxy
Cherry-pick: a6922c068 for 7.1.3 branch only

The ~StoredValueProxy path may dereference it's non owning StoredValue*
and this is a source of bugs.

1) For ephemeral if the StoredValueProxy::value is marked stale it can
be deleted by the stale-item remover, allowing for a
heap-use-after-free error.

2) A number of HashTable update paths decide to drop the HashTable lock
manually, exposing the StoredValueProxy::value to possible manipulation
before the statistics are finalised - a likely source of subtle
statistic inaccuracies (and is a unlocked read violation).

Change-Id: Ie2d5f2cd7a7caf0db1a6c3ae1658d56422835261
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/182505
Reviewed-by: James H <james.harrison@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>