MB-53922: Remove ~StoredValueProxy
The ~StoredValueProxy path may dereference it's non owning StoredValue*
and this is a source of bugs.

1) For ephemeral if the StoredValueProxy::value is marked stale it can
be deleted by the stale-item remover, allowing for a
heap-use-after-free error.

2) A number of HashTable update paths decide to drop the HashTable lock
manually, exposing the StoredValueProxy::value to possible manipulation
before the statistics are finalised - a likely source of subtle
statistic inaccuracies (and is a unlocked read violation).

Change-Id: I990639cb22e19e6c36dd9614d3d3bf419ae5bc81
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/180825
Well-Formed: Restriction Checker
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>