MB-49071: Make engine-specific storage type-safe [2/2]
Using an interface type for the engine-specific storage allows us to
avoid reinterpret_casts from void* and to use smart pointers for
automatic memory management of objects with lifetime bound by the
command context.

This patch replaces the void*-based engine specific API of ep-engine
with a type-safe alternative and updates all callers to use it.

Change-Id: I008bed200d80d717decf03abc3aa1eba9c7d1f84
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/179617
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>