MB-54087: Respond only once for RangeScan error conditions
KV is sending 2 mcbp response packets to the client if a continue
is terminated by one of the following errors.

* scanned VBucket changes state
* scanned collection is dropped
* scan is cancelled

This issue only occurs if the RangeScan is actively scanning and
detects the error in the scan callback. The bug occurs as the
RangeScanDataHandler first transmits the error response using
Cookie::sendResponse. The scan ends and calls notifyIOComplete(status).
Next the frontend executor re-runs (ewouldblock pattern) and triggers a
second response packet for these error codes.

The fix here is to ensure that for the fatal errors, the
RangeScanDataHandler does not send a response and leaves the error
response to the frontend executor, this also ensures that the cluster
map would be properly included for not-my-vbucket errors along side any
other error context data.

Change-Id: Id217c393da6808523bf5d65a81a34fbe973a55c0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181126
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>