MB-53920: CheckpointQuota accounts detached checkpoints
Since when introduced, the CheckpointQuota accounts the memory usage of
checkpoints owned by CM. Ie, detached-to-be-destroyed checkpoints
(owned by CheckpointDestroyer) aren't accounted.

The problem with that approach is that in the case of slow checkpoint
deallocation (for whatever reason), the checkpoint allocation in
Destroyers can take up to the entire BucketQuota.

With this patch we change CheckpointQuota to accounting mem-alloc in
Destroyers too. The consequence is that the system will enter a TempOOM
phase when checkpoint allocation is high (usual CheckpointQuota
rules apply), regardless of what's owning those checkpoints.

Note that this change has implications in the way we perform memory
recovery from CM. By that:

  a. If the memory releasable by detached-to-be-destroyed checkpoints
     allows to push the overall checkpoint alloc down to the lower_mark
     (which is our usual target), then the CheckpointMemRecoveryTask
     isn't woken up at all
  b. Otherwise, the CheckpointMemRecoveryTask is scheduled to run as
     usual, but now it attempts mem-releasing from CM only to the extent
     of what needs to be freed on top of detached-to-be-destroyed
     checkpoints

That is mainly for avoiding excessive ItemExpel/CursorDrop on CM, when
CM isn't the problematic component.

Note: This change might possibly make visible any suboptimal behaviour
on checkpoint mem-releasing in Destroyers. In case, that is expected to
show up as the system entering TempOOM phases in a state where the
Destroyer-alloc component is prevalent compared to CM-alloc.
Regardless of that, optimizations in the Destroyer deallocation area
are already in-progress in MB-53531.

Change-Id: I8540b7781b0929ea6006e1522ee3fe35f19d93ef
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181149
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Paolo Cocchi <paolo.cocchi@couchbase.com>