MB-53953: Cancel all RangeScans on shutdown
If the backend=magma it will block if snapshots are still referenced
when requested to shutdown. Any RangeScan that exists will be in this
state and if not already cancelled or continued will cause a deadlock
on the shutdown.

This commit requests that all RangeScans be cancelled when the bucket
is shutting down.

Change-Id: Ic0ff1867269590a740f3d93dc6fe9f8a29792c78
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181151
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>