MB-53914: Disable PauseResumeTest/PauseFailsWhenPaused while investigating
This test is currently intermittently failing. The issue is how the
test is written, and the fact that Pause is a non-blocking (async)
request means there's no guarantee of what order Pause commands are
executed with respect to other Pause or Resume commands.

This is because KV returns immediately (with “Success”) as soon as the
Pause request is accepted, and then schedules the actual Pause
operation on a background task as it can take potentially a long time
to quiesce disk. Then, if another resume request (or a second, “extra”
Pause request) comes in, that could be executed before the first
Pause.

What we see in the test is the following sequence of operations:

1. Pause request, wait for Bucket to change state to Paused.
2. Pause request, which should do nothing as Bucket is already Paused.
3. Resume request, to verify we can correctly resume.

Because the bulk of the Pause work is done in a background task, (2)
is not actually guaranteed to occur before (3) - indeed we can see (3)
get executed first causing the Bucket to resume, then (2) is scheduled
to run by the executor pool and pauses the Bucket again. This is why
at the end of the test, SelectBucket fails, as the bucket is Paused
and cannot be selected.

Disabling this test while we discuss how to deal with this (should the
API be changed?)

Change-Id: I58c28b62ac357e47ea558a8c329ca96a44c90ea3
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/181160
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>